pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meshcore-bot"
version = "0.1.0"
description = "BLE MeshCore companion-node channel responder bot"
requires-python = ">=3.10"
dependencies = [
  "meshcore==2.2.14",
  "PyYAML==6.0.2",
]

[project.scripts]
meshcore-setup = "meshcore_bot.setup_ble:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
